moved CloudfrontFactsServiceManager to module_utils/aws as a common class
